Consistently handle TimeLineID as UInt32 everywhere
TimeLineID is an uint32 in PostgreSQL but since PostgreSQL's
type system is lacking unsigned types there was no consistent
mode of transmission in the streaming replication protocol and
the timeline got transmitted both as integer and as bigint.
As a consequence although we realized the TimeLineID's unsignedness
we also had heterogenity regarding timeline id's ocurrences across
Npgsql's replication API.
AS of PG 16 the backend is consistently transmitting TimeLineID as
bigint that (due to being backed by an uint32 field) can only have
values >= 0.
With this commit we accept TimeLineID as both integer and bigint
and expose it as uint (UInt32) in all our public APIs.

Closes #5046

+++ src/Npgsql/Replication/PhysicalReplicationSlot.cs
@@ -17,7 +17,7 @@ public class PhysicalReplicationSlot : ReplicationSlot
     /// <param name="slotName">The name of the existing replication slot</param>
     /// <param name="restartLsn">The replication slot's <c>restart_lsn</c></param>
     /// <param name="restartTimeline">The timeline ID associated to <c>restart_lsn</c>, following the current timeline history.</param>
-    public PhysicalReplicationSlot(string slotName, NpgsqlLogSequenceNumber? restartLsn = null, ulong? restartTimeline = null)
+    public PhysicalReplicationSlot(string slotName, NpgsqlLogSequenceNumber? restartLsn = null, uint? restartTimeline = null)
         : base(slotName)
     {
         RestartLsn = restartLsn;
@@ -32,5 +32,5 @@ public class PhysicalReplicationSlot : ReplicationSlot
     /// <summary>
     /// The timeline ID associated to <c>restart_lsn</c>, following the current timeline history.
     /// </summary>
-    public ulong? RestartTimeline { get; }
+    public uint? RestartTimeline { get; }
 }
